refactor(frontend): split SearchaliciousSearchMixin into smaller mixins
Decoupled the monolithic frontend search controller mixin into:
- SearchaliciousStateMixin: handles state, properties and signal updates
- SearchaliciousDomMixin: handles DOM querying (facets, charts, sorts)
- SearchaliciousApiMixin: handles search API fetching and params building

Composed them back in search-ctl.ts. Validated with TypeScript build and passing tests.
